删除部分md文件,重构apis文件内容.使导入变得优雅.移除部分未使用的三方库.加快install速度.
