+++ config/default.ts
@@ -4,7 +4,6 @@ import { Settings as ProSettings } from '@ant-design/pro-layout';
 
 export default {
   navTheme: 'light',
-  // 拂晓蓝
   primaryColor: '#1890ff',
   layout: 'mix',
   contentWidth: 'Fluid',
