--- src/apis/auth.ts
@@ -1,9 +1,0 @@
-/**
- * 验证功能相关
- */
-export const AuthApi = {
-  /**
-   * 登录API
-   */
-  login: 'Identity/LoginAsync'
-};
