--- src/apis/menu.ts
@@ -1,19 +1,0 @@
-/**
- * 菜单相关
- */
-export const MenuApi = {
-  getAuthorityTree: 'Menu/GetTreeAsync',
-  getAuthorityPageData: 'Menu/GetTableAsync',
-  /**
-   * 新增用户
-   */
-  addUser: '',
-  /**
-   * 修改用户
-   */
-  editUser: '',
-  /**
-   * 删除用户
-   */
-  deleteUser: 'User/Delete'
-};
