--- src/apis/role.ts
@@ -1,10 +1,0 @@
-/**
- * 权限相关
- */
-export const RoleApi = {
-  getRolePage: 'Role/GetRolePageAsync',
-  addRole: 'Role/CreateAsync',
-  editRole: 'Role/UpdateAsync',
-  deleteRole: 'Role/Delete',
-  selectRole: 'Role/GetRoleSelectListAsync'
-};
