--- src/apis/user.ts
@@ -1,10 +1,0 @@
-/**
- * 用户相关
- */
-export const UserApi = {
-  getPageUser: 'User/GetUserPageAsync',
-  addOrUpdate: 'User/AddOrUpdateAsync',
-  editUser: '',
-  deleteUser: 'User/DeleteAsync',
-  loadUser: 'User/LoadAsync'
-};
