--- src/assets/readme.md
@@ -1,3 +1,0 @@
-# assets 文件夹
-
-用于存放一些静态资源,如图片,或者其他媒体文件
