+++ src/components/IconFont/index.tsx
@@ -4,11 +4,11 @@ import { createFromIconfontCN } from '@ant-design/icons';
 interface IconFontProps extends IconBaseProps {
   type: string;
 }
+
+const scriptUrl: string = '//at.alicdn.com/t/font_1709014_wl3qyrmbhuo.js';
 /**
  * 使用IconFont字体图标
  */
-const IconFont: React.FC<IconFontProps> = createFromIconfontCN({
-  scriptUrl: '//at.alicdn.com/t/font_1709014_wl3qyrmbhuo.js'
-});
+const IconFont: React.FC<IconFontProps> = createFromIconfontCN({ scriptUrl });
 
 export default IconFont;
