--- src/components/readme.md
@@ -1,3 +1,0 @@
-# components 文件夹
-
-用于存放自定义 React 组件.
