--- src/configs/readme.md
@@ -1,3 +1,0 @@
-# configs 文件夹
-
-用于存放全局的一些固定的配置.
