--- src/locales/readme.md
@@ -1,3 +1,0 @@
-# locales 文件夹
-
-用于存放全球化的一些配置.
