--- src/pages/readme.md
@@ -1,3 +1,0 @@
-# pages 文件夹
-
-Umi 约定文件夹,用于存放页面组件文件
