+++ src/typings/global.d.ts
@@ -11,20 +11,3 @@ declare module '*.gif';
 declare module '*.bmp';
 declare module '*.tiff';
 declare module 'omit.js';
-
-// google analytics interface
-interface GAFieldsObject {
-  eventCategory: string;
-  eventAction: string;
-  eventLabel?: string;
-  eventValue?: number;
-  nonInteraction?: boolean;
-}
-interface Window {
-  ga: (command: 'send', hitType: 'event' | 'pageview', fieldsObject: GAFieldsObject | string) => void;
-  reloadAuthorized: () => void;
-}
-
-declare let ga: Function;
-
-declare const REACT_APP_ENV: 'test' | 'dev' | 'pre' | false;
