--- src/utils/readme.md
@@ -1,3 +1,0 @@
-# utils 文件夹
-
-用于存放一些工具类.工具函数,以及全局统一的一些配置组件.
